requirements.txt
chardet==3.0.4
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.3
django-hosts==2.0
gunicorn==19.6.0
idna==2.5
lxml==3.8.0
nltk==3.2.4
numpy==1.13.1
pandas==0.20.3
psycopg2==2.7.3
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.2
six==1.10.0
urllib3==1.22
pytest==3.2.1
pytest-django==3.1.2
factory-boy==2.9.2
Faker==0.7.18
mock==2.0.0



